Change to sdk style for easy CI/CD
