+++ .gitignore
@@ -110,3 +110,5 @@ Generated_Code #added for RIA/Silverlight projects
 _UpgradeReport_Files/
 Backup*/
 UpgradeLog*.XML
+/.vs
+/src/.vs/PDRating
