--- RatingView.nuspec
@@ -1,25 +1,0 @@
-<?xml version="1.0"?>
-<package xmlns="http://schemas.microsoft.com/packaging/2012/06/nuspec.xsd">
-  <metadata>
-    <id>RatingView</id>
-    <version>2.0.2</version>
-    <title>RatingView</title>
-    <authors>Adam Patridge,RatingView contributors</authors>
-    <owners>Adam Patridge</owners>
-    <licenseUrl>https://github.com/patridge/PDRating/blob/master/LICENSE</licenseUrl>
-    <projectUrl>https://github.com/patridge/PDRating/</projectUrl>
-    <iconUrl>https://github.com/patridge/PDRating/raw/master/component/icons/pdrating_64x64.png</iconUrl>
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <description>Display and optionally collect user ratings for items in your Xamarin.iOS app using a custom rating image, custom scale, and an easily composible transparent background.</description>
-    <releaseNotes>
-v2.0.2: Corrected namespace (breaking, but nothing internal changed).
-v2.0.1: Updated component and samples to support Xamarin.iOS Unified API (again, after breaking changes in the latest Unified API release).
-v2.0.0: Updated component and samples to support Xamarin.iOS Unified API.
-  </releaseNotes>
-    <copyright>© 2013-2018 Adam Patridge</copyright>
-    <tags>xamarin ios rating review</tags>
-  </metadata>
-  <files>
-    <file src="src/PDRating/bin/unified/Release/PDRating.dll" target="lib/Xamarin.iOS10/PDRating.dll" />
-  </files>
-</package>
