--- component.yaml
@@ -1,31 +1,0 @@
-version: 2.0.2
-name: PDReview
-id: pdreview
-publisher: Adam Patridge
-license: component/License.md
-details: component/Details.md
-getting-started: component/GettingStarted.md
-summary: PDRatingView lets you display an average rating and, optionally, collect a user's rating submission on items in your application.
-icons:
-  - component/icons/pdrating_128x128.png
-  - component/icons/pdrating_512x512.png
-publisher-url: https://github.com/patridge/PDRating
-build:
-  - src/PDRating.sln
-libraries:
-  ios: src/PDRating/bin/classic/Release/PDRating.dll
-  ios-unified: src/PDRating/bin/unified/Release/PDRating.dll
-screenshots:
-  - component/Screenshots/five-stars-scale.png
-  - component/Screenshots/six-moustaches-scale.png
-  - component/Screenshots/ten-tomatoes-scale.png
-  - component/Screenshots/custom-background.png
-samples:
-  - name: "iOS Sample"
-    path: samples/PDRatingSample.sln
-    configuration: "Release|iPhone"
-    nobuild: true
-  - name: "iOS Classic Sample"
-    path: samples/PDRatingSample-Classic.sln
-    configuration: "Release|iPhone"
-    nobuild: true
