+++ src/PDRating/PDRating-Classic.csproj
@@ -43,6 +43,5 @@
     <Compile Include="UIImageExtensions.cs" />
     <Compile Include="SizeFExtensions.cs" />
   </ItemGroup>
-  <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
   <Import Project="$(MSBuildExtensionsPath)\Xamarin\iOS\Xamarin.MonoTouch.CSharp.targets" />
 </Project>
