+++ src/PDRating/SizeFExtensions.cs
@@ -1,24 +1,8 @@
 using System;
-
-#if __UNIFIED__
-using UIKit;
-using Foundation;
 using CoreGraphics;
-#else
-using MonoTouch.UIKit;
-using MonoTouch.Foundation;
-using MonoTouch.CoreGraphics;
 
-using System.Drawing;
-using CGRect = global::System.Drawing.RectangleF;
-using CGPoint = global::System.Drawing.PointF;
-using CGSize = global::System.Drawing.SizeF;
-using nfloat = global::System.Single;
-using nint = global::System.Int32;
-using nuint = global::System.UInt32;
-#endif
-
-namespace PatridgeDev {
+namespace PatridgeDev
+{
     public static class SizeFExtensions {
         public static CGSize ScaleProportional(this CGSize original, nfloat maxWidth, nfloat maxHeight) {
             nfloat ratioX = (float)maxWidth / original.Width;
