+++ src/PDRating/UIImageExtensions.cs
@@ -1,20 +1,4 @@
-#if __UNIFIED__
 using UIKit;
-using Foundation;
-using CoreGraphics;
-#else
-using MonoTouch.UIKit;
-using MonoTouch.Foundation;
-using MonoTouch.CoreGraphics;
-
-using System.Drawing;
-using CGRect = global::System.Drawing.RectangleF;
-using CGPoint = global::System.Drawing.PointF;
-using CGSize = global::System.Drawing.SizeF;
-using nfloat = global::System.Single;
-using nint = global::System.Int32;
-using nuint = global::System.UInt32;
-#endif
 
 namespace PatridgeDev {
     public static class UIImageExtensions {
