+++ src/PDRatingSample/Info.plist
