+++ src/PDRatingSample/Resources/Icon-60@2x.png
