+++ src/PDRatingSample/Resources/Icon-72.png
