+++ src/PDRatingSample/Resources/Icon-72@2x.png
