+++ src/PDRatingSample/Resources/Icon-76.png
