+++ src/PDRatingSample/Resources/Icon-76@2x.png
