+++ src/PDRatingSample/Resources/Icon.png
