+++ src/PDRatingSample/Resources/Icon@2x.png
