+++ src/PDRatingSample/Resources/Moustaches/chosen.png
