+++ src/PDRatingSample/Resources/Moustaches/empty.png
