+++ src/PDRatingSample/Resources/Moustaches/filled.png
