+++ src/PDRatingSample/Resources/Stars/chosen.png
