+++ src/PDRatingSample/Resources/Stars/empty.png
