+++ src/PDRatingSample/Resources/Stars/filled.png
