+++ src/PDRatingSample/Resources/Tomatoes/chosen.png
